+++ modules/protocol/src/testing/utils/transfer.ts
@@ -8,6 +8,7 @@ import {
   ResolveTransferParams,
   TransferResolver,
   DEFAULT_TRANSFER_TIMEOUT,
+  Balance,
 } from "@connext/vector-types";
 import {
   createlockHash,
@@ -18,7 +19,7 @@ import {
   expect,
 } from "@connext/vector-utils";
 import { defaultAbiCoder } from "@ethersproject/abi";
-import { BigNumberish } from "@ethersproject/bignumber";
+import { BigNumber, BigNumberish } from "@ethersproject/bignumber";
 import { AddressZero } from "@ethersproject/constants";
 
 import { env } from "../env";
@@ -26,23 +27,26 @@ import { chainId } from "../constants";
 
 // Will create a hashlock transfer in the channel, and return the full
 // transfer state (including the necessary resolver)
-// TODO: Should be improved to create any type of state, though maybe
-// this is out of scope for integration test utils
 export const createTransfer = async (
   channelAddress: string,
-  payor: IVectorProtocol,
-  payee: IVectorProtocol,
+  creator: IVectorProtocol,
+  resolver: IVectorProtocol,
   assetId: string = AddressZero,
   amount: BigNumberish = 10,
+  outsiderPayee?: string,
+  channelInitialBalance?: Balance,
+  skipBalanceVerification = false, // use true if testing concurrency
 ): Promise<{ channel: FullChannelState; transfer: FullTransferState }> => {
   // Create the transfer information
   const preImage = getRandomBytes32();
   const lockHash = createlockHash(preImage);
+  const payorAddress = creator.signerAddress;
+  const payeeAddress = outsiderPayee ?? resolver.signerAddress;
+
   const balance = {
-    to: [payor.signerAddress, payee.signerAddress],
+    to: [payorAddress, payeeAddress],
     amount: [amount.toString(), "0"],
   };
-
   const transferInitialState = createTestHashlockTransferState({ lockHash });
   const params: CreateTransferParams = {
     channelAddress,
@@ -54,13 +58,20 @@ export const createTransfer = async (
     assetId,
   };
 
-  const ret = await payor.create(params);
+  const preCreateChannel = await creator.getChannelState(channelAddress);
+  const assetIdx = (preCreateChannel?.assetIds ?? []).findIndex((a) => a === assetId);
+  const preCreateBalance = channelInitialBalance ?? preCreateChannel!.balances[assetIdx];
+  const isAlice = creator.signerAddress === preCreateChannel?.alice;
+  const initCreatorBalance = preCreateBalance.amount[isAlice ? 0 : 1];
+  const initResolverBalance = preCreateBalance.amount[isAlice ? 1 : 0];
+
+  const ret = await creator.create(params);
   expect(ret.getError()).to.be.undefined;
   const channel = ret.getValue();
-  expect(await payee.getChannelState(channelAddress)).to.be.deep.eq(channel);
+  expect(await resolver.getChannelState(channelAddress)).to.be.deep.eq(channel);
 
   const { transferId } = (channel.latestUpdate as ChannelUpdate<typeof UpdateType.create>).details;
-  const transfer = await payee.getTransferState(transferId);
+  const transfer = await resolver.getTransferState(transferId);
   expect(transfer).to.containSubset({
     balance,
     assetId,
@@ -81,6 +92,15 @@ export const createTransfer = async (
   const decoded = defaultAbiCoder.decode([transfer!.transferEncodings[1]], encoding)[0];
   expect(decoded.preImage).to.be.deep.eq(preImage);
   expect(transfer!.transferEncodings.length).to.be.eq(2);
+
+  // Ensure the balance was properly decremented for creator && not touched for
+  // resolver
+  if (!skipBalanceVerification) {
+    const finalCreatorBalance = channel?.balances[assetIdx].amount[isAlice ? 0 : 1];
+    const finalResolverBalance = channel?.balances[assetIdx].amount[isAlice ? 1 : 0];
+    expect(BigNumber.from(finalCreatorBalance)).to.be.eq(BigNumber.from(initCreatorBalance).sub(balance.amount[0]));
+    expect(BigNumber.from(finalResolverBalance)).to.be.eq(BigNumber.from(initResolverBalance).sub(balance.amount[1]));
+  }
 
   return {
     channel,